alerter/requirements.txt
flask==2.0.1
requests==2.26.0
pyyaml==6.0
python-dotenv==0.19.2
aiosmtplib==1.1.6
jinja2==3.0.3
gunicorn==20.1.0 